Initial Content Push
This commit contains all the initial content built for this project.
All work was done prior to the creation of this repo (meaning there is
no revision history before this point). Project contains all types of
sprites and a basic setup. Should be stable, but not guaranteed error
free. Biggest future direction from here is to comment the Blueprints as
there currently are none.